Refactor: remove calculatable module, replace Scoreboard class methods with instance methods, fix specs
