+++ bowling_scores.rb
@@ -1,20 +1,16 @@
 require 'optparse'
 require './modules/validatable'
 require './modules/checkable'
-require './modules/calculable'
 require './modules/printable'
 require './classes/scoreboard'
 require 'byebug'
 require 'i18n'
-I18n.config.available_locales = :es
-I18n.locale = :es
-I18n.load_path = ['config/locales/es.yml']
-I18n.backend.load_translations
+require './config/i18n'
 
 OptionParser.new do |parser|
   parser.banner = 'Usage: bowling-scores.rb [options]'
   parser.on('-f', '--filename FILE_NAME', 'The name of the scores text file (with extension).') do |filename|
-    ScoreBoard.game_score(filename)
+    ScoreBoard.new(filename).game_score
   end
   parser.on('-h', '--help', 'Show this help message') do
     puts parser
