+++ config/i18n.rb
@@ -1,0 +1,4 @@
+I18n.config.available_locales = :es
+I18n.locale = :es
+I18n.load_path = ['config/locales/es.yml']
+I18n.backend.load_translations
